Exercicio de casting modulo 9 correção

+++ TarefaCastingMod9/src/TarefaCasting.java
@@ -1,10 +1,10 @@
 public class TarefaCasting {
     public static void main(String[] args) {
-        
+
         int numero = 10;
         System.out.println(numero);
 
-        long numeroMaior =  numero;
+        Integer numeroMaior =  numero;
         System.out.println(numeroMaior);
 
     }
